requirements.txt
aiogram>=3.6
fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
psycopg[binary,pool]>=3.1
alembic>=1.13
httpx>=0.27
pydantic>=2.7
pydantic-settings>=2.4
python-dotenv>=1.0
PyJWT>=2.9
pytz>=2024.1   # опц.; в коді використано zoneinfo